small changes for checkall option, thin walls and point model is perfectly positioned.
made couple changes to the 'checkall' option. note: the checkall process still needs a bit of work for it to be a 1 stop test to see what ER roles need adjusting.
thinwalls should be a 'better' test to run compared to last versions.
point model is now perfectly positioned in xy relative to the numbers it sits between.